% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{external_demo_data}
\alias{external_demo_data}
\title{Data for biomass demonstrations}
\format{
A dataframe with 15 rows and 12 columns:
\describe{
\item{watershed}{Whatershed in which the data were collected}
\item{year}{Year in which the data were collected}
\item{plot}{Plot in which the individual tree was measured}
\item{elev_m}{Elevation in meters above sea level}
\item{hli}{Heat load index}
\item{steep_deg}{Slope in degrees}
\item{exp_factor}{Expansion factor, number of trees per hectare that a given plot tree represents}
\item{species}{Species of the individual tree, using four-letter species codes}
\item{status}{Live or Dead}
\item{vigor}{0-3 for live trees. 4 or 5 for dead trees.}
\item{dbh_cm}{Diameter at breast height in cm}
\item{cbh_m}{Crown base height in meters}
}
}
\source{
Created by Kea Rutherford for demonstration purposes
}
\usage{
external_demo_data
}
\description{
A fake dataset created for biomass demonstration purposes only
}
\keyword{datasets}
