% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_checks.R
\name{HBEFBiomass}
\alias{HBEFBiomass}
\title{HBEFBiomass}
\usage{
HBEFBiomass(data_type = "internal", external_data, results = "by_plot")
}
\arguments{
\item{data_type}{describe...}

\item{external_data}{describe...}

\item{results}{describe...}
}
\value{
describe...
}
\description{
Describe...
}
