% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubbardBrookForestAnalytics-package.R
\docType{package}
\name{HubbardBrookForestAnalytics-package}
\alias{HubbardBrookForestAnalytics}
\alias{HubbardBrookForestAnalytics-package}
\title{HubbardBrookForestAnalytics: Compute tree biomass for Hubbard Brook Experimental Forest}
\description{
Facilitates easy access and consistent application of the revised Hubbard Brook Experimental Forest tree biomass equations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kearutherford/HubbardBrookForestAnalytics}
  \item Report bugs at \url{https://github.com/kearutherford/HubbardBrookForestAnalytics/issues}
}

}
\author{
\strong{Maintainer}: Kea Rutherford \email{krutherford@berkeley.edu}

Authors:
\itemize{
  \item John Battles
}

}
\keyword{internal}
